requirements.txt
instabot==0.26.0
Pillow==5.4.1
python-dotenv==0.10.1
requests==2.21.0
python-dotenv===0.10.1